moves types and interfaces to specific types file
